Official Release (c) v1.0.0 - The Release Update
Over the course of 3 months, I've really had fun creating this extension with the others.
It's an honour to be able to use this as a project for learning about creating Chrome extensions, as well as many other things such as getting better at JavaScript.

I hope you enjoy this extension - as there's been hours upon hours poured into it and I hope you get good use out of it. Can't wait to see you all in the future!

Signing off,
Solar (and the team)