Manifest + Do 2/3 suggestions in #65 (Pin Notices, Removed Loader) (v1.0.1-dev)
Pin notices functionality has been added, although I'm not sure if the day thing works yet, I'll have to check.

Removed loader because it was deemed unnecessary and it would take time to implement another loader.

Updated manifest for a better description and buffed version to v1.0.1 to prepare for doing so.